+++ config/app.php
@@ -114,6 +114,8 @@ return [
 
     'providers' => [
 
+        Dingo\Api\Provider\LaravelServiceProvider::class,
+        
         /*
          * Laravel Framework Service Providers...
          */